checkout

+++ src/Routes/PatientController.js
@@ -268,6 +268,29 @@ const getListMed = async (req, res) => {
 }
 
 router.get('/getAllMedicine', getListMed);
+
+router.get('/cartinCheckOut', async (req, res) => {
+  let exists = await PatientModel.findById(req.user);
+  if (!exists || req.user.__t!=="Patient") {
+    return res.status(500).json({
+      success: false,
+      message: "Not authorized"
+    });
+  }  let pid=req.user._id//temp until login
+  const cartItems = await CartItem.find({userId:pid});
+  let list=[]
+  for(var x in cartItems){
+    const med=await MedicineModel.findById(cartItems[x].medicineId);
+    list.push(med);
+  }
+  Result={
+    cartItems:cartItems,
+    medInfo:list
+  }
+  console.log(cartItems);
+  res.json(Result);
+});
+
 // Delete an item from the cart
 router.delete('/deleteCartItem/:cartItemId',protect, async (req, res) => {
   let exists = await PatientModel.findById(req.user);
@@ -392,14 +415,22 @@ router.get('/getAllMedicine', protect, async (req, res) => {
 // Checkout and create an order
 router.post('/checkout', async (req, res) => {
   try {
+    let exists = await PatientModel.findById(req.user);
+    if (!exists || req.user.__t != "Patient") {
+      return res.status(500).json({
+        success: false,
+        message: "Not authorized"
+      });
+    }
+    const patientId=req.user._id;
     // Get the items from the cart
-    const cartItems = await CartItem.find();
+    const cartItems = await CartItem.find({userId: patientId});
     console.log(cartItems);
 
     // Initialize variables for order creation
     let total = 0;
     const itemsForOrder = [];
-
+7
     // Calculate the total cost and construct the order
     for (const cartItem of cartItems) {
       const medicine = await MedicineModel.findById(cartItem.medicineId);
@@ -435,13 +466,15 @@ router.post('/checkout', async (req, res) => {
 
     // Create the order in the database
     const newOrder = new Order({
+      userId:patientId,
       items: itemsForOrder,
-      total,
+      total:total,
+      address: req.body.address,
     });
     await newOrder.save();
 
     // Clear the cart by removing all cart items
-    await CartItem.deleteMany();
+    await CartItem.deleteMany({userId: patientId});
 
     res.json({ message: 'Checkout successful. Your order has been placed.' });
   } catch (error) {
@@ -451,8 +484,18 @@ router.post('/checkout', async (req, res) => {
 });
 
 // Add a new delivery address for a patient
+
 router.post('/addAddress', async (req, res) => {
-  const { patientId, newAddress } = req.body; // Assuming you send patientId and newAddress in the request body
+  let exists = await PatientModel.findById(req.user);
+    if (!exists || req.user.__t != "Patient") {
+      return res.status(500).json({
+        success: false,
+        message: "Not authorized"
+      });
+    }
+    const patientId=req.user._id;//temp untill login
+  const newAddress=req.body.address;
+ // const { patientId, newAddress} = req.body; // Assuming you send patientId and newAddress in the request body
   console.log(newAddress);
   try {
     // Find the patient by their ID
@@ -481,8 +524,12 @@ router.post('/addAddress', async (req, res) => {
     patient.address.push(newAddress);
     console.log(patient.address);
     // Save the updated patient data
-    const updatedPatient = await patient.save();
-
+   // const updatedPatient = await patient.save();
+    const updatedPatient = await PatientModel.findOneAndUpdate({ _id: patientId },
+      {
+          address: patient.address,
+          
+      });
     res.json({ message: 'Address added successfully', patient: updatedPatient });
   } catch (error) {
     console.error(error);
@@ -492,8 +539,16 @@ router.post('/addAddress', async (req, res) => {
 
 
 // Retrieve the patient's addresses
-router.get('/getAddresses/:patientId', async (req, res) => {
-  const patientId = req.params.patientId;
+router.get('/getAddresses', async (req, res) => {
+  let exists = await PatientModel.findById(req.user);
+    if (!exists || req.user.__t != "Patient") {
+      return res.status(500).json({
+        success: false,
+        message: "Not authorized"
+      });
+    }
+    const patientId = req.user._id;
+
 
   try {
     // Find the patient by their ID
